+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+### Added
+
+- `compliance` form element
+- `hint` property to form elements with labels
+
 ## 1.15.0 (2021-03-03)
 
 ### Added